terraPorta_api/requirements/requirements.txt
django==2.1
djangorestframework==3.9.0
django-redis-cache==1.8.1
djangorestframework-jwt==1.11.0
Markdown==3.0.1
psycopg2-binary==2.7.6.1
celery==4.2.1
redis==2.10.6
django-filter==2.0.0
pyjwt==1.7.0
cryptography==2.4.2
stripe==2.16.0
django-cors-headers==2.4.0
uWSGI==2.0.17.1
whitenoise==4.1.2
